refactor: コードをモジュール化する (#20)
